Allow related in addition to established connections
We want certain related packages to be accepted. In particular,
I've run into issues using a VPN in front of a Whonix Gateway. This
because the related Fragmentation Needed packages get dropped:

 1780 280.196298  10.137.0.31 → 10.137.0.19  ICMP 590 Destination unreachable (Fragmentation needed)

With those packages not being delivered, a connection can be
established and works fine until a large package is sent. However,
once that happens no more packages can be delivered and the
connection times out after a while.

There is probably also a bunch of other, related packages that
we want to be delivered. ICMP host unreachable comes to mind, for
instance. Without it being delivered, one has to wait for
the connection attempt to timeout.

Note: Related connections appear to have been disabled in
414c2105149e02 but with no clear indication to why.